Add event_time query parameter to events endpoint
The events/events endpoint currently has no documented way to indicate
the time range for which data is being requested. This adds an
event_time query parameter to the API endpoint that takes the form
of an ISO 8601 UTC hour YYYY-MM-DDTHH so that the user can request
all events that occurred within that specific hour. This is the same
format used for MDS events [1].

[1]: https://github.com/openmobilityfoundation/mobility-data-specification/tree/2.0.1/provider#historical-events---query-parameters