Add Curb Event field for operator device name
When working with partners we have frequently found that they have
device identifiers that are not UUIDs that we would like them to include
in their curb event data for easy reference, but the CDS spec does not
currently have a field for them. This adds a data_source_device_name
field to the Curb Event spec to hold such data. It is expected to be
provided in all cases in which the provider has such an identifier.
This is quite analagous to the user_zone_id field on the Curb Zone
object [1].

[1]: https://github.com/openmobilityfoundation/curb-data-specification/tree/1.0.0/curbs#curb-zone